monitor-data-pipeline/collect-agent/requirements.txt
# Collect Agent Requirements

# gRPC for receiving metrics from compute node agents
grpcio>=1.60.0,<1.62.0
grpcio-tools>=1.60.0,<1.62.0
protobuf>=4.21.6,<5.0.0

# etcd client for configuration (use etcd3-py fork for protobuf 4.x compatibility)
etcd3-py>=0.1.6

# Kafka for message broker
aiokafka>=0.8.0
kafka-python>=2.0.2
